Introduce the RCW compiler
Introduce the RCW compiler, which is a Python program that converts a
human-readable "source" representation of the RCW and generates an RCW
binary (with PBL header/footer and CRC) from it.

The RCW source files are intended to be a replacement for the binhex text
files that are typically used to generate RCW binaries.

File p4080ds/R_PPSXX_0xe/rcw_2sgmii_1500mhz.rcw is included as an example.
To avoid confusion, the compiler is not used during a normal build.

Signed-off-by: Timur Tabi <timur@freescale.com>