rcw.py: use an ordered dictionary for the symbols
The RCW compiler uses Python dictionaries to store symbols and assignments,
but dictionaries are unordered.  So when using the -r option to generate
an rcw source file, the symbols appear to be in random order, instead of
following the order in the SOC definition file.

So we use a custom "ordered dictionary" class to allow the symbols to
be written in the same order they were read.

Signed-off-by: Timur Tabi <timur@freescale.com>