p4080ds: add note explaining how to enable RGMII with R_RRRSS_0x16
The P4080 RCW R_RRRSS_0x16 enables support for RGMII on FM1@DTSEC2, but it
won't work unless U-Boot disables Lane B on Bank 3.  This bank cannot be
disabled in the RCW itself because of erratum SERDES_8.  To avoid confusion,
add a note explaining this requirement in the RCW text file.

Signed-off-by: Timur Tabi <timur@freescale.com>