Remove copyright messages from RCWs
RCW files are just numbers -- a collection of bits that configure the SOC.
They have no creative content in them, therefore, they cannot be copyrighted.

Claiming that RCWs can be copyrighted implies that RCWs can be given
restrictive licenses.  This would allow someone to create an RCW for a
common configuration and prevent others from using that configuration.

Signed-off-by: Timur Tabi <timur@freescale.com>