rcw.py: use a comma to separate PBI parameters
Use a comma to separate the first and second parameters in PBI commands
that take two parameters.  This will make it possible to support
expressions in parameters, because we now use a comma instead of a blank
space to deliniate them.

Unfortunately, this requires changing most of the source files to conform
to the next syntax.

Signed-off-by: Timur Tabi <timur@freescale.com>