rcw.py: add -D option to specify bitfield values on the command line
This lets you define a new value for a bitfield on the command-line, so
that you can generate quick derivative RCWs without editing text files.

Usage:
	rcw.py ... -D <bitfield>=<value> ...

Repeat -D for each bitfield.

Example:

	rcw.py -D MEM_PLL_RAT=12 -i source.rcw -o rcw.bin

This will compile source.rcw, but MEM_PLL_RAT is set to 12 instead of
whatever is in source.rcw.

Signed-off-by: Timur Tabi <timur@freescale.com>