T4240: fix FMan frequency
After reworking the board:
1. Replace R2145 with 140 ohm resistor (value was 180 ohms).
Resistor size is 0402.
2. Replace R2530 with 0 ohm resistor (value was 10 ohms).
Resistor size is 0402.

MDIO issue is fixed when using 4 XAUI riser cards, so we fix the FMan
frequency to align the requirement of workaround A-006064:

1. For SerDes1 or SerDes2 protocol configurations (via SRDS_PRTCL_S1,
SRDS_PRTCL_S2 RCW bits) that include XFI or XAUI, the associated FMAN
must be configured to operate in the range of 450MHz to 466.667MHz.

2. For SerDes1 or SerDes2 protocol configurations that do not include
XFI or XAUI (leaving QSGMII, 1G SGMII, and 2.5G SGMII), the associated
FMAN must be configured to operate in the range of 333MHz to 350MHz.

For rcw_1_1_6_6, both FMan1 & FMan2 are configured to 467MHz.
For rcw_1_28_6_12, configure FMan1 to 467MHz, configure FMan2 to 350MHz.

Signed-off-by: Shaohui Xie <Shaohui.Xie@freescale.com>