P4080DS: add rcw for PCIe end point support
To configure the PCIe1 and PCIe2 as agent mode using serdes protocol 5.

The SerDes lane E/F/G/H of Bank1 is assigned to PCIe 2 and is routed to
Slot 3, the link width is x4 and link speed is 5G.

Ther SerDes lane A/B/C/D of Bank1 is assigned to PCIe 1 and lane A/B is
routed to Slot 1, lane C/D is routed to Slot 2. For slot 1 and slot 2,
the link width is x2 and link speed is also 5G.

Signed-off-by: Mingkai Hu <Mingkai.Hu@freescale.com>
Change-Id: Ib3bd51fcee9764887a9df877788b1778d34ca0e7
Reviewed-on: http://git.am.freescale.net:8181/1595
Reviewed-by: Zang Tiefei-R61911 <tie-fei.zang@freescale.com>
Reviewed-by: Schmitt Richard-B43082 <B43082@freescale.com>
Tested-by: Schmitt Richard-B43082 <B43082@freescale.com>