B4860: Slow MDC clock to comply IEEE specs
With the default value of MDIO_CLK_DIV generatee MDC is too high and It is
violating the IEEE specs much higher than 2.5Mhz.

Although there is errata(A-006260) for EMI2(MDIO2), but same errata is
been hit on EMI1(MDIO1). unfortunately this errata never hit on B4 rev1.
So reduced the MDIO_CLK_DIV value.

Signed-off-by: Prabhakar Kushwaha <prabhakar@freescale.com>
Change-Id: Ic98ff774c0a7ee356f45dfef84aad398e1222286
Reviewed-on: http://git.am.freescale.net:8181/3550
Reviewed-by: Leekha Shaveta-B20052 <shaveta@freescale.com>
Reviewed-by: Aggrwal Poonam-B10812 <Poonam.Aggrwal@freescale.com>
Reviewed-by: Schmitt Richard-B43082 <B43082@freescale.com>
Tested-by: Schmitt Richard-B43082 <B43082@freescale.com>