t4240: create RCW files for rev2.0
Mostly, Rev2.0 RCWs have same settings as Rev1.0, with differences as
below:

1. MEM_PLL_RAT uses a value which is half of Rev1.0, it's 12 in Rev1.0,
for Rev2.0 it uses 6.

2. removed all errata workarounds which fixed in Rev2.0, the errata list
as below:

IFC bus speed work-around
A-005600
A-006031
A-005147

Since the IFC erratum is fixed, the PBI_SRC is set to 16 bit NOR.
set PBI_SRC=14.

Also, slow down the MDC clock to make it <= 2.5 MHZ.

Signed-off-by: Shaohui Xie <Shaohui.Xie@freescale.com>
Change-Id: I4abd6f28796169c20a97574a2d6778090710637b
Reviewed-on: http://git.am.freescale.net:8181/6116
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Tiefei Zang <roy.zang@freescale.com>
Reviewed-by: Lijun Pan <Lijun.Pan@freescale.com>