P4080DS: add rcw for PCIe RC X4 link and 5G support
To configure the PCIe1 and PCIe2 as RC mode using serdes protocol 5.
The SerDes lane E/F/G/H of Bank1 is assigned to PCIe 2 and is routed
to Slot 3, the link width is x4 and link speed is 5G.
The SerDes laneA/B/C/D of Bank1 is assigned to PCIe 1 and lane A/B is
routed to Slot 1, lane C/D is routed to Slot 2. For slot 1 and slot 2,
the link width is x2 and link speed is also 5G.

Signed-off-by: Mingkai Hu <Mingkai.Hu@freescale.com>
Change-Id: I7990a9be6c18abec532b4b1fdfb7b499c84d9ed0
Reviewed-on: http://git.am.freescale.net:8181/11866
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Lijun Pan <Lijun.Pan@freescale.com>