arm: ls102xa: Adjust the PBI command about the load address of U-Boot for NOR boot
The original load address of U-Boot is 0x67f80000. The address
space of NOR flash is 0x60000000 to 0x67ffffff. It will cause
the size of u-boot couldn't be larger than 512K. As more features
are supported in u-boot, the size of u-boot is larger than 512K.

To fix this issue, the PBI command about the load address of U-Boot
for NOR boot is adjusted to 0x60100000.

Signed-off-by: Alison Wang <alison.wang@freescale.com>
Change-Id: If9eedf618e682447edb1e20d7367ff99e59e46f7
Reviewed-on: http://git.am.freescale.net:8181/36036
Reviewed-by: Zhengxiong Jin <Jason.Jin@freescale.com>
Tested-by: Zhengxiong Jin <Jason.Jin@freescale.com>