Replace old byte_swap.tcl infrastructure for QSPI booting
This patch makes all qspiboot RCW binaries be automatically generated as
properly byte-swapped directly by the Python script.

This is possible via the %littleendian64b=1 option.

The byte_swap.tcl script had a hack that did not swap the last 8 bytes
of the binary provided at input.  This is because all subsequent PBI
commands after a <qspi_endianness.rcw> are also endianness-reversed,
including the "Stop and CRC" command.  Not swapping the final command is
the workaround so that its endianness is correct in the end.

But there are also some more byte_swap.tcl scripts in ls1021aqds/ and in
ls1043aqds/.  These do not skip the last 8 bytes from swapping.
But after commit 1a2389a28a13 (Revert "qspi: use byte_swap.tcl under
board directory for special board"), these scripts are effectively
left-overs, and no one is using them.  So simply remove them.

To match byte_swap.tcl's behavior of not swapping the last 8 bytes, the
%dont64bswapcrc=1 option is used.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>