lx2160a: Add workaround for erratum A-050234
The elastic buffer integrated in the receiver path of PCI Express
controller is used to compensate the worst-case frequency difference
between link partners on each side of the PCI Express link by adding
or removing SKP Symbols (part of SKIP Ordered Sets) toward the
received serial data stream.

Due to this erratum, the default configuration of the elastic buffer
thresholds in the PCI Express controller is not optimal and can cause
various link errors.

Signed-off-by: Hou Zhiqiang <Zhiqiang.Hou@nxp.com>