ls1046a: offer SerDes1 protocol 0x3333 with PLL mapping 1111
The LS1046A reset state machine is a bit limited in the SRDS_PRTCL_S1
protocol combinations that it offers. For example, it offers protocol
0x3333 (4x SGMII) only with PLL mapping 2222, and that is only
sufficient as long as dynamic protocol switching between 1G (SGMII) and
10G (XFI) isn't desired at runtime.

If that is desired as well, then we need PLL2 (yes, specifically 2) to
be provisioned with a refclk of 156.25 MHz, since this is what XFI needs.
But that is in conflict with the PLL mapping of the aforementioned native
0x3333 RCW protocol, because 1G (SGMII) can only work with a PLL refclk
of 100 or 125 MHz, so that's what has to be fed into PLL 2 with the
native 0x3333 option.

Dynamic frequency switching of PLLs is hard and out of question.
There are fewer PLLs shared by multiple SerDes lanes by design of this
SerDes block family, and when one lane wants its PLL reconfigured to
produce a different clock net, it is virtually impossible to guarantee
in the general sense that other lanes, using the same PLL, do not, in
fact, want that PLL to stay at the same clock net. So the best we can do
is keep the 2 PLLs read-only and just map lanes to one or the other.

It is desirable to be able to use PLL2 for the 156.25 MHz refclk
frequency (for the 10G link modes), and PLL1 for the 100 MHz refclk
frequency (for the 1G link modes). It turns out, this is possible, even
if not 100% natively through the reset state machine built-in options.

The strategy is to pick a pair of SerDes refclk frequencies in the RCW
that is correct for the given board (thus allowing the SerDes PLLs to
lock, and the SoC to boot) as a first step. The SerDes protocol can be
absolutely anything as long as the PLL frequencies are right, because as
a second step, we'll be fixing up the SerDes lane registers for the
desired final protocol, to appear as if we had that protocol natively.

Applied to this case, "absolutely anything" means we start with the
native SerDes protocol of 0x1133. To reach protocol 0x3333, we then need
to reconfigure SerDes lanes C and D to SGMII.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>