lx2160aqds_rev2: extend SD1 protocol 13 with 100G MAC 1 split into 4 x 25G ports
The SD1 protocol 13 defines 2 100G ports which will span all the 8 lanes
on this SerDes block. What this patch does is to enable the usecase in
which the SerDes block #1 is configured to provide 4 x 25G ports and one
100G port.

The newly added rcw_2000_700_2600_13_3_2_e100g1_split.rcw file is based
on rcw_2000_700_2600_13_3_2.rcw. The only difference is that the
e100g1_split.rcw is used to reconfigure LANES E-H (that were used by
MAC.1) to run independently as 4 x 25G ports.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>
Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>