ls1028a: make all switch PCSes addressable on the MDIO bus
On LS1028A, the PCS blocks which go to the 4 external Ethernet switch ports
are all located on the same internal MDIO bus of the switch. The SerDes
SGMIIaCR1[MDEV_PORT] controls their address on the bus, and by default,
it is zero for all devices. To make them individually addressable, we need
to map the SGMII PCS of switch port i to MDIO address i.

This is already the established convention in Linux drivers, which expect
this mapping to have already been applied, as well as in U-Boot, which fixes
up MDEV_PORT from within the switch driver to apply this mapping.

Add PBL sequences for all SerDes lanes which use SGMII or 2500Base-X
to write SGMIIaCR1[MDEV_PORT] on those respective lanes. This makes the
U-Boot fixup unnecessary, which is useful for example when the U-Boot
switch driver doesn't run.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>