lx2160a: fix SerDes lane documentation
The table with RCW[SRDS_PRTCL_S1] values in LX21060ARM.pdf and
LX2162ARM.pdf is wrong. Lane 0 is not the left-most lane, but actually
the right-most one. The letters are what's correct (where 'A' would
correspond to 0, 'H' to 7).

Yet, this project describes the lanes as if reading the SRDS_PRTCL_S1
table from left to right, but that's reading the lanes from 7 to 0.
Make sure the README clearly specifies that, to avoid confusion.

I don't think changing RCW file paths is an option at this stage.
There might be build tools which have hardcoded the existing paths.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>