rcw: ls1028a: Add configuration for audio codec usage
Add a tweaked rcw_1500_gpu600 configuration with EC1_SAI4_5_PMUX = 2.
This switches the EC1_TXD2 pin to SAI4_TX_DATA mode and EC1_GTX_CLK to
SAI4_TX_SYNC mode (and a few others) so that we can communicate with the
STGL5000 audio codec.

Note: Dip switch SW5_8 needs to be set to on for the external mux to
      route these signals to the STGL5000 rather than to the IEEE1588
      connector(s).

Signed-off-by: Martyn Welch <martyn.welch@collabora.com>